+++ index.html
@@ -10,9 +10,9 @@
 <head>
     <meta charset="utf-8" name="viewport" content="inital-scale=1, width=device-width">
     <title>seman-game</title>
-    <script type="text/javascript" src="/jquery/js/jquery-3.5.1.js"></script>
-    <script type="text/javascript" src="/scripts/script.js"></script>
-    <link rel="stylesheet" href="/templates/style.css">
+    <script type="text/javascript" src="./static/js/jquery/js/jquery-3.5.1.js"></script>
+    <script type="text/javascript" src="./static/js/script.js"></script>
+    <link rel="stylesheet" href="./static/css/style.css">
 </head>
     <body>
         <h1>Bienvenue dans la ligue des légendes</h1>
