+++ static/js/jquery/css/ui-lightness/jquery-ui-1.10.0.custom.css
